changed int64_t to double for kmeans

+++ src/machine_learning/clustering/kmeans/kmeans.cc
@@ -4,23 +4,23 @@
 
 int main(){
   srand((unsigned)time(NULL));
-  std::vector<std::vector<int64_t> > data;
+  std::vector<std::vector<double> > data;
   for(int i = 0; i<100; i++){
-    int rand1 = (rand() % 5) + 0;
-    int rand2 = (rand() % 5) + 0;
+    double rand1 = (rand() % 5) + 0;
+    double rand2 = (rand() % 5) + 0;
     data.push_back({rand1, rand2});
   }
 
   for(int i = 0; i<100; i++){
-    int rand1 = (rand() % 50) + 40;
-    int rand2 = (rand() % 50) + 40;
+    double rand1 = (rand() % 50) + 40;
+    double rand2 = (rand() % 50) + 40;
     data.push_back({rand1, rand2});
   }
 
   json j;
   j["data"] = data;
   kmeans a(data, 2);
-  std::pair<std::vector<std::vector<int64_t> >, std::map<std::vector<int64_t> ,int64_t> > ans = a.run();
+  std::pair<std::vector<std::vector<double> >, std::map<std::vector<double> ,int64_t> > ans = a.run();
   j["centroids"] = ans.first;
   j["assignments"] = ans.second;
   std::ofstream file("info.json");
